+++ test/smoke/testclient_test.go
@@ -68,15 +68,25 @@ func (c *TestClient) Configure(server, dockerReg string) error {
 	return nil
 }
 
-func (c *TestClient) Cmd(t *testing.T, args ...string) *exec.Cmd {
+// allArgs produces a []string representing all args determined by the sous
+// subcommand, sous flags and any other args.
+func allArgs(subcmd string, f *sousFlags, args []string) []string {
+	allArgs := strings.Split(subcmd, " ")
+	allArgs = append(allArgs, f.Args()...)
+	allArgs = append(allArgs, args...)
+	return allArgs
+}
+
+func (c *TestClient) Cmd(t *testing.T, subcmd string, f *sousFlags, args ...string) *exec.Cmd {
 	t.Helper()
-	cmd := mkCMD(c.Dir, c.BinPath, args...)
+	cmd := mkCMD(c.Dir, c.BinPath, allArgs(subcmd, f, args)...)
 	cmd.Env = append(cmd.Env, fmt.Sprintf("SOUS_CONFIG_DIR=%s", c.ConfigDir))
 	return cmd
 }
 
-func (c *TestClient) Run(t *testing.T, args ...string) (string, error) {
-	cmd := c.Cmd(t, args...)
+func (c *TestClient) Run(t *testing.T, subcmd string, f *sousFlags, args ...string) (string, error) {
+	t.Helper()
+	cmd := c.Cmd(t, subcmd, f, args...)
 	fmt.Fprintf(os.Stderr, "SOUS_CONFIG_DIR = %q\n", c.ConfigDir)
 	fmt.Fprintf(os.Stderr, "running sous in %q: %s\n", c.Dir, args)
 	// Add quotes to args with spaces for printing.
@@ -88,7 +98,7 @@ func (c *TestClient) Run(t *testing.T, args ...string) (string, error) {
 	out := &bytes.Buffer{}
 	cmd.Stdout = io.MultiWriter(os.Stdout, out)
 	cmd.Stderr = os.Stderr
-	fmt.Fprintf(os.Stderr, "==> sous %s\n", strings.Join(args, " "))
+	fmt.Fprintf(os.Stderr, "==> sous %s\n", strings.Join(allArgs(subcmd, f, args), " "))
 	err := cmd.Run()
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
@@ -96,19 +106,47 @@ func (c *TestClient) Run(t *testing.T, args ...string) (string, error) {
 	return out.String(), err
 }
 
-func (c *TestClient) MustRun(t *testing.T, args ...string) string {
+func (c *TestClient) MustRun(t *testing.T, subcmd string, f *sousFlags, args ...string) string {
 	t.Helper()
-	out, err := c.Run(t, args...)
+	out, err := c.Run(t, subcmd, f, args...)
 	if err != nil {
 		t.Fatal(err)
 	}
 	return out
 }
 
-func (c *TestClient) MustFail(t *testing.T, args ...string) {
+func (c *TestClient) MustFail(t *testing.T, subcmd string, f *sousFlags, args ...string) {
 	t.Helper()
-	_, err := c.Run(t, args...)
+	_, err := c.Run(t, subcmd, f, args...)
 	if err == nil {
 		t.Fatalf("command should have failed: sous %s", args)
+	}
+}
+
+func (c *TestClient) TransformManifestAsString(t *testing.T, getSetFlags *sousFlags, f func(manifest string) string) {
+	manifest := c.MustRun(t, "manifest get", getSetFlags)
+	manifest = f(manifest)
+	manifestSetCmd := c.Cmd(t, "manifest set", getSetFlags)
+	manifestSetCmd.Stdin = ioutil.NopCloser(bytes.NewReader([]byte(manifest)))
+	if out, err := manifestSetCmd.CombinedOutput(); err != nil {
+		t.Fatalf("manifest set failed: %s; output:\n%s", err, out)
+	}
+}
+
+func (c *TestClient) TransformManifest(t *testing.T, getSetFlags *sousFlags, f func(m sous.Manifest) sous.Manifest) {
+	manifest := c.MustRun(t, "manifest get", getSetFlags)
+	var m sous.Manifest
+	if err := yaml.Unmarshal([]byte(manifest), &m); err != nil {
+		t.Fatalf("manifest get returned invalid YAML: %s\nInvalid YAML was:\n%s", err, manifest)
+	}
+	m = f(m)
+	manifestBytes, err := yaml.Marshal(m)
+	if err != nil {
+		t.Fatalf("failed to marshal updated manifest: %s\nInvalid manifest was:\n% #v", err, m)
+	}
+	manifestSetCmd := c.Cmd(t, "manifest set", getSetFlags)
+	manifestSetCmd.Stdin = ioutil.NopCloser(bytes.NewReader(manifestBytes))
+	if out, err := manifestSetCmd.CombinedOutput(); err != nil {
+		t.Fatalf("manifest set failed: %s; output:\n%s", err, out)
 	}
 }